+++ src/lib-index/mail-transaction-log.c
@@ -58,7 +58,8 @@ static void mail_transaction_log_2_unlink_old(struct mail_transaction_log *log)
 		return;
 	}
 
-	if (st.st_mtime + MAIL_TRANSACTION_LOG2_STALE_SECS <= ioloop_time) {
+	if (st.st_mtime + MAIL_TRANSACTION_LOG2_STALE_SECS <= ioloop_time &&
+	    !log->index->readonly) {
 		if (unlink(log->filepath2) < 0 && errno != ENOENT) {
 			mail_index_set_error(log->index,
 				"unlink(%s) failed: %m", log->filepath2);
